+++ src/test/java/com/imsweb/mph/Mph2018RuleTests.java
@@ -578,6 +578,15 @@ public class Mph2018RuleTests {
         Assert.assertEquals(ruleCountToTest, output.getAppliedRules().size());
         Assert.assertEquals(ruleStepToTest, output.getStep());
         Assert.assertTrue(output.getReason().contains("different rows in Table 3"));
+        //table 2 and table 3
+        i1.setHistologyIcdO3("8540");
+        i1.setBehaviorIcdO3("2");
+        i2.setHistologyIcdO3("8523");
+        output = _utils.computePrimaries(i1, i2);
+        Assert.assertEquals(MphUtils.MpResult.MULTIPLE_PRIMARIES, output.getResult());
+        Assert.assertEquals(ruleCountToTest, output.getAppliedRules().size());
+        Assert.assertEquals(ruleStepToTest, output.getStep());
+        Assert.assertTrue(output.getReason().contains("different rows in Table 3"));
 
         // Rule M15 Abstract a single primary (the invasive) when an in situ tumor is diagnosed after an invasive tumor in the same breast.
         ruleStepToTest = "M15";
@@ -2237,6 +2246,16 @@ public class Mph2018RuleTests {
         Assert.assertEquals(ruleStepToTest, output.getStep());
         i1.setHistologyIcdO3("8333");
         i2.setHistologyIcdO3("8083");
+        output = _utils.computePrimaries(i1, i2);
+        Assert.assertEquals(MphUtils.MpResult.MULTIPLE_PRIMARIES, output.getResult());
+        Assert.assertEquals(ruleCountToTest, output.getAppliedRules().size());
+        Assert.assertTrue(output.getReason().contains("two or more different subtypes/variants in Column 3, Table 3"));
+        Assert.assertEquals(ruleStepToTest, output.getStep());
+        //
+        i1.setHistologyIcdO3("8250");
+        i1.setBehaviorIcdO3("2");
+        i2.setHistologyIcdO3("8256");
+        i2.setBehaviorIcdO3("3");
         output = _utils.computePrimaries(i1, i2);
         Assert.assertEquals(MphUtils.MpResult.MULTIPLE_PRIMARIES, output.getResult());
         Assert.assertEquals(ruleCountToTest, output.getAppliedRules().size());