+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Multiple Primary and Histology Coding Rules Version History
 **Changes in version 1.20**
 
-- Updated the readme file.
+- Fixed a bug in 2018 breast and lung rules.
 
 **Changes in version 1.19**